Vagrant::Helpers::Ssh module, remove Net::SSH remnants, tests passing
